Organize the adapter and holder for the game
